[AArch64] match TRN starting from undef elements
When the first element of a trn mask is undef, the isTRNMask function
assumes that the value for "WhichResult" should be 1. That has a 50%
chance of being wrong, so we fail to match some valid trn1/trn2.
This patch introduces a more precise test to determine the correct
value of "WhichResult", based on corresponding code in the isZIPMask
and isUZPMask functions.